+++ src/plugins/add-query-output.js
@@ -146,10 +146,11 @@ class AddQueryOutputToViewModel {
 
   /**
    * Queries for related documents based on similar tags and searches the storage provider.
+   * @template T The viewModel we are manipulating.
    * @param {string} eventLabel The event label to run queries for.
-   * @param {object} viewModel A Uttori view-model object.
+   * @param {T} viewModel A Uttori view-model object.
    * @param {import('../../dist/custom.js').UttoriContext} context A Uttori-like context.
-   * @returns {Promise<object | object[]>} The provided view-model document.
+   * @returns {Promise<T>} The provided view-model document.
    * @example <caption>AddQueryOutputToViewModel.callback(viewModel, context)</caption>
    * const context = {
    *   config: {