+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.1.0](https://github.com/uttori/uttori-wiki/compare/v6.0.3...v6.0.4) - 2024-01-06
+
+- 🧰 Add support for redirecting via Express route syntax to help migrate away from Jekyll
+- 🧰 Export `asyncHandler` for use in plugins or other projects
+- 🛠 Clean up some types that were still generic
+- 🎁 Update dev dependencies
+
 ## [6.0.4](https://github.com/uttori/uttori-wiki/compare/v6.0.3...v6.0.4) - 2023-12-25
 
 - 🛠 Ensure `tags` are always sorted